contrib: Add U-Boot Sandbox board configuration
Added a board configuration, for building and running the U-Boot
Sandbox binary on the lab host. It serves as a handy, easy-to-run
testbed instead of real hardware.

Also included is a local lab-and-builder board configuration which
can be used to build and run the the U-Boot Sandbox on the local
machine. Additionally args/sandbox-local contains the arguments
which needs to be passed to tbot to make this run out-of-the-box.

A tweak required to the subprocess channel for the Sandbox to work
is also included.

Signed-off-by: Niel Fourie <lusus@denx.de>